Swagger para os controllers Book e Auth

+++ src/main/java/vitordev/project/projetoSpring/service/PersonServices.java
@@ -55,11 +55,19 @@ public class PersonServices implements UserDetailsService {
     }
 
     public void deleteAll(){
+        if(repository.findAll().isEmpty()){
+            throw new ResourceNotFoundException("No records found to delete!");
+        }
+
         repository.deleteAll();
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("No records found for this ID!");
+        }
+
+        repository.deleteById(id);;
     }
 }
